==> Scripts/16.Conformer generator.ipynb <==
from rdkit import Chem
import os
import pandas as pd

folder = "../Data/other/SDF poses"
data = []

for file in os.listdir(folder):
    if file.endswith(".sdf"):
        base_name = file[:-4]  # remove '.sdf'
        if "_VINA" in base_name:
            continue
        suppl = Chem.SDMolSupplier(os.path.join(folder, file))
        for i, mol in enumerate(suppl):
            if mol:
                smiles = Chem.MolToSmiles(mol)
                data.append({"name": base_name, "smiles": smiles})

df = pd.DataFrame(data)

#save to csv

import pandas as pd
import mygene

papyrus = pd.read_csv("../Data/resources/subset/papyrus_subset.tsv", sep="\t")
acvr2b_smiles = papyrus.loc[papyrus["accession"] == "Q13705", "SMILES"].unique()
papyrus_filtered = papyrus[~papyrus["SMILES"].isin(acvr2b_smiles)]

papyrus_random = papyrus_filtered.sample(n=45, random_state=42)[["accession", "SMILES", "InChIKey"]]
papyrus_random.rename(columns={"SMILES": "smiles"}, inplace=True)

mg = mygene.MyGeneInfo()
res = mg.querymany(papyrus_random['accession'].tolist(), scopes='uniprot', fields='symbol', species='human')
df_genes = pd.DataFrame(res)[['query', 'symbol']].rename(columns={'query': 'accession'})

papyrus_random = papyrus_random.merge(df_genes, on='accession', how='left')\
    .drop_duplicates()\
    .dropna(subset=['symbol'])\
    .reset_index(drop=True)

papyrus_random['name'] = papyrus_random['InChIKey'].str[:4] + '_' + papyrus_random['symbol']

papyrus_random = papyrus_random.drop(columns=['accession', 'InChIKey', 'symbol'])

papyrus_random


import pandas as pd

combined_df = pd.concat([df, papyrus_random], ignore_index=True)
combined_df

combined_df.to_csv('../Data/docs/7.compounds_for_validaton.csv', index=False)

import pandas as pd
from rdkit import Chem
from rdkit.Chem import AllChem
import os

#first remove old isntances of the sdf file
if os.path.exists("all_conformers.sdf"):
    os.remove("all_conformers.sdf")

df = pd.read_csv("../Data/docs/7.compounds_for_validaton.csv")
writer = Chem.SDWriter("../Data/other/all_conformers.sdf")

for idx, row in df.iterrows():
    mol = Chem.MolFromSmiles(row['smiles'])
    mol = Chem.AddHs(mol)
    mol.SetProp("_Name", str(row['name']))
    ids = AllChem.EmbedMultipleConfs(mol, numConfs=50, randomSeed=42)
    AllChem.UFFOptimizeMoleculeConfs(mol)
    
    for conf_id in ids:
        writer.write(mol, confId=conf_id)

writer.close()

import pandas as pd
from rdkit import Chem
from dimorphite_dl import run_with_mol

df = pd.read_csv("../Data/docs/6.compounds for vina smiles.csv")
writer = Chem.SDWriter("all_protomers.sdf")

for idx, row in df.iterrows():
    mol = Chem.MolFromSmiles(row['SMILES'])
    protomers = run_with_mol(mol, pH=7.4, min_ph=6.0, max_ph=8.0)

    for protomer in protomers:
        protomer = Chem.AddHs(protomer)
        Chem.EmbedMolecule(protomer)
        writer.write(protomer)

writer.close()
